+++ tests/SphinxQL/FacetTest.php
@@ -29,12 +29,11 @@ class FacetTest  extends \PHPUnit\Framework\TestCase
             'title' => 'what is there to do', 'content' => 'we need to create dummy data for tests'),
     );
 
-    public static function setUpBeforeClass()
-    {
-        $conn = TestUtil::getConnectionDriver();
-        $conn->setParam('port', 9307);
-        self::$conn = $conn;
-    }
+    public static function setUpBeforeClass(): void{
+		$conn = TestUtil::getConnectionDriver();
+		$conn->setParam('port', 9307);
+		self::$conn = $conn;
+	}
 
     /**
      * @return Facet