+++ tests/SphinxQL/SphinxQLTest.php
@@ -3,7 +3,7 @@
 use Foolz\SphinxQL\Expression;
 use Foolz\SphinxQL\Facet;
 use Foolz\SphinxQL\Helper;
-use Foolz\SphinxQL\Match;
+use Foolz\SphinxQL\MatchBuilder;
 use Foolz\SphinxQL\SphinxQL;
 use Foolz\SphinxQL\Tests\TestUtil;
 
@@ -30,7 +30,7 @@ class SphinxQLTest extends \PHPUnit\Framework\TestCase
             'title' => 'what is there to do', 'content' => 'we need to create dummy data for tests'),
     );
 
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         $conn = TestUtil::getConnectionDriver();
         $conn->setParam('port', 9307);
@@ -573,7 +573,7 @@ class SphinxQLTest extends \PHPUnit\Framework\TestCase
 
         $this->assertCount(2, $result);
 
-        $match = (new Match($this->createSphinxQL()))
+        $match = (new MatchBuilder($this->createSphinxQL()))
             ->field('content')
             ->match('directly')
             ->orMatch('lazy');